Solved the Android bug
Solved the android bug (camera size needs to be fixed to world
coordinates, not device coordinates). The Polygon Sprite Batch for
some reason needs to be flushed between two polygon sprites.

Also separated the image selection from the catwalk. The code is still
a mess, but this will be necessary for loading multiple levels.