Added Level Progression and 3 levels.
Changed the Gameplay scene loading method to take a "level class" as a
parameter.  Level classes have indication for the number of enemies,
and the files to load as a background and foreground.  Levels will
advance when completed, and repeat when done.

If you repeat the level too many times, I have encountered a bug, but
I haven't been able to reproduce.